// types/testimonial.ts
export interface Testimonial {
  id: number
  name: string
  role: string
  content: string
  image: string
  rating: number
  date: string
  learned: string
  valuable: string
  coping: string
  strategies: string
  aid: string
  made: string
  meaningful: string
  connections: string
  others: string
  on: string
  similar: string
  journeys: string
}